# server/ui/api/v1/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "OpenAPI 2.0 documentation for [Mailpit](https://github.com/axllent/mailpit).",
    "title": "Mailpit API",
    "contact": {
      "name": "GitHub",
      "url": "https://github.com/axllent/mailpit"
    },
    "license": {
      "name": "MIT license",
      "url": "https://github.com/axllent/mailpit/blob/develop/LICENSE"
    },
    "version": "v1"
  },
  "paths": {
    "/api/v1/info": {
      "get": {
        "description": "Returns basic runtime information, message totals and latest release version.",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "application"
        ],
        "summary": "Get application information",
        "operationId": "AppInformation",
        "responses": {
          "200": {
            "$ref": "#/responses/AppInfoResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/message/{ID}": {
      "get": {
        "description": "Returns the summary of a message, marking the message as read.\n\nThe ID can be set to `latest` to return the latest message.",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "message"
        ],
        "summary": "Get message summary",
        "operationId": "GetMessageParams",
        "parameters": [
          {
            "type": "string",
            "example": "4oRBnPtCXgAqZniRhzLNmS",
            "description": "Message database ID or \"latest\"",
            "name": "ID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Message",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/responses/NotFoundResponse"
          }
        }
      }
    },
    "/api/v1/message/{ID}/headers": {
      "get": {
        "description": "Returns the message headers as an array. Note that header keys are returned alphabetically.\n\nThe ID can be set to `latest` to return the latest message headers.",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "message"
        ],
        "summary": "Get message headers",
        "operationId": "GetHeadersParams",
        "parameters": [
          {
            "type": "string",
            "example": "4oRBnPtCXgAqZniRhzLNmS",
            "description": "Message database ID or \"latest\"",
            "name": "ID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "MessageHeadersResponse",
            "schema": {
              "$ref": "#/definitions/MessageHeadersResponse"
            }
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/responses/NotFoundResponse"
          }
        }
      }
    },
...
    "/api/v1/message/{ID}/part/{PartID}": {
      "get": {
        "description": "This will return the attachment part using the appropriate Content-Type.\n\nThe ID can be set to `latest` to reference the latest message.",
        "produces": [
          "application/*",
          "image/*",
          "text/*"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "message"
        ],
        "summary": "Get message attachment",
        "operationId": "AttachmentParams",
        "parameters": [
          {
            "type": "string",
            "example": "4oRBnPtCXgAqZniRhzLNmS",
            "description": "Message database ID or \"latest\"",
            "name": "ID",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "example": "2",
            "description": "Attachment part ID",
            "name": "PartID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/BinaryResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/responses/NotFoundResponse"
          }
        }
      }
    },
...
    "/api/v1/message/{ID}/raw": {
      "get": {
        "description": "Returns the full email source as plain text.\n\nThe ID can be set to `latest` to return the latest message source.",
        "produces": [
          "text/plain"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "message"
        ],
        "summary": "Get message source",
        "operationId": "DownloadRawParams",
        "parameters": [
          {
            "type": "string",
            "example": "4oRBnPtCXgAqZniRhzLNmS",
            "description": "Message database ID or \"latest\"",
            "name": "ID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/TextResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/responses/NotFoundResponse"
          }
        }
      }
    },
...
    "/api/v1/webui": {
      "get": {
        "description": "Returns configuration settings for the web UI.\nIntended for web UI only!",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "application"
        ],
        "summary": "Get web UI configuration",
        "operationId": "WebUIConfiguration",
        "responses": {
          "200": {
            "$ref": "#/responses/WebUIConfigurationResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Address": {
      "description": "An address such as \"Barry Gibbs \u003cbg@example.com\u003e\" is represented\nas Address{Name: \"Barry Gibbs\", Address: \"bg@example.com\"}.",
      "type": "object",
      "title": "Address represents a single mail address.",
      "properties": {
        "Address": {
          "type": "string"
        },
        "Name": {
          "type": "string"
        }
      },
      "x-go-package": "net/mail"
    },
    "AppInformation": {
      "description": "AppInformation struct",
      "type": "object",
      "properties": {
        "Database": {
          "description": "Database path",
          "type": "string"
        },
        "DatabaseSize": {
          "description": "Database size in bytes",
          "type": "number",
          "format": "double"
        },
        "LatestVersion": {
          "description": "Latest Mailpit version",
          "type": "string"
        },
        "Messages": {
          "description": "Total number of messages in the database",
          "type": "number",
          "format": "double"
        },
        "RuntimeStats": {
          "description": "Runtime statistics",
          "type": "object",
          "properties": {
            "Memory": {
              "description": "Current memory usage in bytes",
              "type": "integer",
              "format": "uint64"
            },
            "MessagesDeleted": {
              "description": "Database runtime messages deleted",
              "type": "number",
              "format": "double"
            },
            "SMTPAccepted": {
              "description": "Accepted runtime SMTP messages",
              "type": "number",
              "format": "double"
            },
            "SMTPAcceptedSize": {
              "description": "Total runtime accepted messages size in bytes",
              "type": "number",
              "format": "double"
            },
            "SMTPIgnored": {
              "description": "Ignored runtime SMTP messages (when using --ignore-duplicate-ids)",
              "type": "number",
              "format": "double"
            },
            "SMTPRejected": {
              "description": "Rejected runtime SMTP messages",
              "type": "number",
              "format": "double"
            },
            "Uptime": {
              "description": "Mailpit server uptime in seconds",
              "type": "number",
              "format": "double"
            }
          }
        },
        "Tags": {
          "description": "Tags and message totals per tag",
          "type": "object",
          "additionalProperties": {
            "type": "integer",
            "format": "int64"
          }
        },
        "Unread": {
          "description": "Total number of messages in the database",
          "type": "number",
          "format": "double"
        },
        "Version": {
          "description": "Current Mailpit version",
          "type": "string"
        }
      },
      "x-go-package": "github.com/axllent/mailpit/internal/stats"
    },
    "Attachment": {
      "description": "Attachment struct for inline and attachments",
      "type": "object",
      "properties": {
        "ContentID": {
          "description": "Content ID",
          "type": "string"
        },
        "ContentType": {
          "description": "Content type",
          "type": "string"
        },
        "FileName": {
          "description": "File name",
          "type": "string"
        },
        "PartID": {
          "description": "Attachment part ID",
          "type": "string"
        },
        "Size": {
          "description": "Size in bytes",
          "type": "number",
          "format": "double"
        }
      },
      "x-go-package": "github.com/axllent/mailpit/internal/storage"
    },
...
    "Message": {
      "description": "Message data excluding physical attachments",
      "type": "object",
      "properties": {
        "Attachments": {
          "description": "Message attachments",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Attachment"
          }
        },
        "Bcc": {
          "description": "Bcc addresses",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Address"
          }
        },
        "Cc": {
          "description": "Cc addresses",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Address"
          }
        },
        "Date": {
          "description": "Message date if set, else date received",
          "type": "string",
          "format": "date-time"
        },
        "From": {
          "$ref": "#/definitions/Address"
        },
        "HTML": {
          "description": "Message body HTML",
          "type": "string"
        },
        "ID": {
          "description": "Database ID",
          "type": "string"
        },
        "Inline": {
          "description": "Inline message attachments",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Attachment"
          }
        },
        "MessageID": {
          "description": "Message ID",
          "type": "string"
        },
        "ReplyTo": {
          "description": "ReplyTo addresses",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Address"
          }
        },
        "ReturnPath": {
          "description": "Return-Path",
          "type": "string"
        },
        "Size": {
          "description": "Message size in bytes",
          "type": "number",
          "format": "double"
        },
        "Subject": {
          "description": "Message subject",
          "type": "string"
        },
        "Tags": {
          "description": "Message tags",
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "Text": {
          "description": "Message body text",
          "type": "string"
        },
        "To": {
          "description": "To addresses",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Address"
          }
        }
      },
      "x-go-package": "github.com/axllent/mailpit/internal/storage"
    },
    "MessageHeadersResponse": {
      "description": "Message headers",
      "type": "object",
      "additionalProperties": {
        "type": "array",
        "items": {
          "type": "string"
        }
      },
      "x-go-name": "messageHeaders",
      "x-go-package": "github.com/axllent/mailpit/server/apiv1"
    },
...
    "WebUIConfiguration": {
      "description": "Response includes global web UI settings",
      "type": "object",
      "properties": {
        "DuplicatesIgnored": {
          "description": "Whether messages with duplicate IDs are ignored",
          "type": "boolean"
        },
        "Label": {
          "description": "Optional label to identify this Mailpit instance",
          "type": "string"
        },
        "MessageRelay": {
          "description": "Message Relay information",
          "type": "object",
          "properties": {
            "AllowedRecipients": {
              "description": "Only allow relaying to these recipients (regex)",
              "type": "string"
            },
            "BlockedRecipients": {
              "description": "Block relaying to these recipients (regex)",
              "type": "string"
            },
            "Enabled": {
              "description": "Whether message relaying (release) is enabled",
              "type": "boolean"
            },
            "ReturnPath": {
              "description": "Enforced Return-Path (if set) for relay bounces",
              "type": "string"
            },
            "SMTPServer": {
              "description": "The configured SMTP server address",
              "type": "string"
            }
          }
        },
        "SpamAssassin": {
          "description": "Whether SpamAssassin is enabled",
          "type": "boolean"
        }
      },
      "x-go-name": "webUIConfiguration",
      "x-go-package": "github.com/axllent/mailpit/server/apiv1"
    }
  },
  "responses": {
    "AppInfoResponse": {
      "description": "Application information",
      "schema": {
        "$ref": "#/definitions/AppInformation"
      }
    },
...
    "BinaryResponse": {
      "description": "Binary data response which inherits the attachment's content type.",
      "schema": {
        "type": "string"
      }
    },
    "ErrorResponse": {
      "description": "Server error will return with a 400 status code\nwith the error message in the body",
      "schema": {
        "type": "string"
      }
    },
...
    "NotFoundResponse": {
      "description": "Not found error will return a 404 status code",
      "schema": {
        "type": "string"
      }
    },
...
    "TextResponse": {
      "description": "Plain text response",
      "schema": {
        "type": "string"
      }
    },
    "WebUIConfigurationResponse": {
      "description": "Web UI configuration response",
      "schema": {
        "$ref": "#/definitions/WebUIConfiguration"
      }
    },
...
  }
}